refactor: export getCrc utility function from principal package
